added code comments, and restructured /client/src to have a pages folder that contains views/pages
