+++ client/src/containers/App.js
@@ -1,13 +1,16 @@
+// dependency imports
 import React from "react";
 import "./App.css";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
-import Home from "../components/home/home";
-import BuySell from "../components/buysellPage/buySellPage";
-import User from "../components/userPage/userPage";
-import Login from "../components/authenticationPages/login";
+// component imports
+import Home from "../components/pages/homePage/home";
+import BuySell from "../components/pages/buysellPage/buySellPage";
+import User from "../components/pages/userPage/userPage";
+import Login from "../components/pages/authenticationPages/login";
 import FootBar from "../components/appWide/footer/footBar";
 
+// class component
 class App extends React.Component {
   render() {
     return (
@@ -15,7 +18,7 @@ class App extends React.Component {
         <header>
           <h1>Hello. I will someday become a video game trader</h1>
         </header>
-        
+          {/* Router will set to a page component. All other components will be placed into those page components */}
           <BrowserRouter>
             <Switch>
               <Route exact path="/" component={Login} />
